Move redirection parser functions
